+++ Tenta/XCTest+Property.swift
@@ -29,27 +29,31 @@ public extension XCTestCase {
 
 public extension XCTestCase {
 
+    func runProperty<TestValue>(_ property: Property<TestValue>, file: StaticString = #file, line: UInt = #line) {
+        if let failedValue = property.checkProperty() {
+            XCTFail("failed with value: \(failedValue), rerun with seed: \(seed)", file: file, line: line)
+        }
+    }
+
     /**
      Run test with specified generator.
      */
     func runTest<TestValue>(
             file: StaticString = #file,
             line: UInt = #line,
-            gen: Generator<TestValue>,
+            generator: Generator<TestValue>,
             predicate: @escaping (TestValue) throws -> Bool
     ) {
 
         let property = Property(
                 description: "testDesc",
-                generator: gen,
+                generator: generator,
                 seed: seed,
                 numberOfTests: numberOfTests,
                 predicate: predicate
         )
 
-        if let failedValue = property.checkProperty() {
-            XCTFail("failed with value: \(failedValue), rerun with seed: \(seed)", file: file, line: line)
-        }
+        runProperty(property, file: file, line: line)
     }
 
     /**
@@ -63,7 +67,7 @@ public extension XCTestCase {
         runTest(
                 file: file,
                 line: line,
-                gen: TestValue.self.generator,
+                generator: TestValue.self.generator,
                 predicate: predicate
         )
     }
@@ -75,11 +79,6 @@ public extension XCTestCase {
             _ secondGenerator: Generator<OtherTestValue>,
             predicate: @escaping (TestValue, OtherTestValue) throws -> Bool
     ) {
-
-//        func unaryPredicate(tuple: (TestValue, OtherTestValue)) throws -> Bool {
-//            return try predicate(tuple.0, tuple.1)
-//        }
-
         let property = Property(
                 description: "",
                 generator: firstGenerator.combine(with: secondGenerator, transform: { ($0, $1) }),
@@ -88,9 +87,7 @@ public extension XCTestCase {
                 predicate: predicate
         )
 
-        if let failedValue = property.checkProperty() {
-            XCTFail("failed with value: \(failedValue), rerun with seed: \(seed)", file: file, line: line)
-        }
+        runProperty(property, file: file, line: line)
     }
 
     func runTest<TestValue: Generatable, OtherTestValue: Generatable>(
@@ -113,47 +110,43 @@ public extension XCTestCase {
     func runWithXCTest<TestValue: Generatable>(
             file: StaticString = #file,
             line: UInt = #line,
-            predicate: @escaping (TestValue) throws -> Void
+            test: @escaping (TestValue) throws -> Void
     ) {
-        runWithXCTest(file: file, line: line, gen: TestValue.generator, predicate: predicate)
+        runWithXCTest(file: file, line: line, generator: TestValue.generator, test: test)
     }
 
     func runWithXCTest<TestValue>(
             file: StaticString = #file,
             line: UInt = #line,
-            gen: Generator<TestValue>,
-            predicate: @escaping (TestValue) throws -> Void
+            generator: Generator<TestValue>,
+            test: @escaping (TestValue) throws -> Void
     ) {
+        let predicate = TestCasePropertyConverter.shared.convert(
+                predicate: test,
+                toBoolPredicate: (),
+                from: self
+        )
 
         let property = Property(
                 description: "",
-                generator: gen,
+                generator: generator,
                 seed: seed,
-                numberOfTests: numberOfTests
-        ) {
-            TestCasePropertyConverter.shared.set(true, for: self)
-            do {
-                try predicate($0)
-            } catch {
-                TestCasePropertyConverter.shared.set(false, for: self)
-            }
-            return TestCasePropertyConverter.shared.passStatus(for: self)
-        }
+                numberOfTests: numberOfTests,
+                predicate: predicate
+        )
 
         TestCasePropertyConverter.shared.set({ _ = property.checkProperty() }, for: self)
 
-        if let failedValue = property.checkProperty() {
-            XCTFail("failed with value: \(failedValue), rerun with seed: \(seed)", file: file, line: line)
-        }
+        runProperty(property, file: file, line: line)
     }
 
     func runWithXCTest<TestValue: Generatable, OtherTestValue: Generatable>(
             file: StaticString = #file,
             line: UInt = #line,
-            predicate: @escaping (TestValue, OtherTestValue) throws -> Void
+            test: @escaping (TestValue, OtherTestValue) throws -> Void
 
     ) {
-        runWithXCTest(file: file, line: line, TestValue.generator, OtherTestValue.generator, predicate: predicate)
+        runWithXCTest(file: file, line: line, TestValue.generator, OtherTestValue.generator, test: test)
     }
 
     func runWithXCTest<TestValue, OtherTestValue>(
@@ -161,27 +154,23 @@ public extension XCTestCase {
             line: UInt = #line,
             _ firstGenerator: Generator<TestValue>,
             _ secondGenerator: Generator<OtherTestValue>,
-            predicate: @escaping (TestValue, OtherTestValue) throws -> Void
+            test: @escaping (TestValue, OtherTestValue) throws -> Void
     ) {
+        let predicate = TestCasePropertyConverter.shared.convert(
+                predicate: test,
+                toBoolPredicate: (),
+                from: self
+        )
         let property = Property(
                 description: "",
                 generator: firstGenerator.combine(with: secondGenerator),
                 seed: seed,
-                numberOfTests: numberOfTests
-        ) {
-            TestCasePropertyConverter.shared.set(true, for: self)
-            do {
-                try predicate($0.0, $0.1)
-            } catch {
-                TestCasePropertyConverter.shared.set(false, for: self)
-            }
-            return TestCasePropertyConverter.shared.passStatus(for: self)
-        }
+                numberOfTests: numberOfTests,
+                predicate: predicate
+        )
 
         TestCasePropertyConverter.shared.set({ _ = property.checkProperty() }, for: self)
 
-        if let failedValue = property.checkProperty() {
-            XCTFail("failed with value: \(failedValue), rerun with seed: \(seed)", file: file, line: line)
-        }
+        runProperty(property, file: file, line: line)
     }
 }